pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "loan-approval"
version = "0.1.0"   
description = "A machine learning model to predict loan approval based on applicant data."
authors = [
    { name = "VVN" }
]
dependencies = [
    "pandas",
    "scikit-learn",
    "joblib",
    "numpy",
    "pyyaml"
]

requires-python = ">=3.9"

[tool.setuptools.packages.find]
where = ["src"]
[project.scripts]
loan-train = "loan_approval.cli:train_cli"
loan-predict = "loan_approval.cli:predict_cli" 